Add serializer for unit test definition
