+++ libs/utils/src/lib/models/unitTest.ts
@@ -1,0 +1,9 @@
+export interface OperationMock {
+  outputs?: string;
+  properties?: Record<string, string>;
+}
+
+export interface Assertion {
+  assertionString: string;
+  description: string;
+}
